requirements.txt
torch
transformers
streamlit
pillow
clip-by-openai
